pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = "Personal & family life assistant"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]
chat = ["anthropic>=0.40"]
calendar = ["google-api-python-client", "google-auth-oauthlib"]
maps = ["googlemaps>=4.10"]

[project.scripts]
assistant = "assistant.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]

[tool.ruff]
line-length = 100
